third_party/WebKit/LayoutTests/imported/wpt/webrtc/rtcpeerconnection/rtcpeerconnection-constructor-expected.txt

This is a testharness.js-based test.
FAIL RTCPeerConnection.length RTCPeerConnection is not defined
FAIL new RTCPeerConnection() RTCPeerConnection is not defined
FAIL new RTCPeerConnection(null) RTCPeerConnection is not defined
FAIL new RTCPeerConnection(undefined) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({}) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: null }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceServers: undefined }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{}] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceServers: [null] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceServers: [undefined] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceServers: [{ urls: "stun:stun1.example.net" }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: [] }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: ["stun:stun1.example.net"] }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: ["stun:stun1.example.net", "stun:stun2.example.net"] }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: "turns:turn.example.org", username: "user", credential: "cred" }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: "turn:turn.example.net", username: "user", credential: "cred" }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: ["turns:turn.example.org", "turn:turn.example.net"], username: "user", credential: "cred" }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: "stun:stun1.example.net", credentialType: "password" }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: "stun:stun1.example.net", credentialType: "token" }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: "turn:turn.example.net" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL new RTCPeerConnection({ iceServers: [{ urls: "turn:turn.example.net", username: "user" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL new RTCPeerConnection({ iceServers: [{ urls: "turn:turn.example.net", credential: "cred" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL new RTCPeerConnection({ iceServers: [{ urls: "turns:turn.example.org" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL new RTCPeerConnection({ iceServers: [{ urls: "turns:turn.example.org", username: "user" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL new RTCPeerConnection({ iceServers: [{ urls: "turns:turn.example.org", credential: "cred" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL new RTCPeerConnection({ iceServers: [{ urls: "relative-url" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" that is not a DOMException SyntaxError: property "code" is equal to undefined, expected 12
FAIL new RTCPeerConnection({ iceServers: [{ urls: "http://example.com" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" that is not a DOMException SyntaxError: property "code" is equal to undefined, expected 12
FAIL new RTCPeerConnection({ iceServers: [{ urls: [], credentialType: "password" }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: [], credentialType: "token" }] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceServers: [{ urls: [], credentialType: "invalid" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceServers: [{ url: "stun:stun1.example.net" }] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceTransportPolicy: null }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceTransportPolicy: undefined }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceTransportPolicy: "relay" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceTransportPolicy: "all" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceTransportPolicy: "invalid" }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceTransportPolicy: "none" }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceTransports: "invalid" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ iceTransports: "none" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ bundlePolicy: null }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ bundlePolicy: undefined }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ bundlePolicy: "balanced" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ bundlePolicy: "max-compat" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ bundlePolicy: "max-bundle" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ bundlePolicy: "invalid" }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ rtcpMuxPolicy: null }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ rtcpMuxPolicy: undefined }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ rtcpMuxPolicy: "negotiate" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ rtcpMuxPolicy: "require" }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ rtcpMuxPolicy: "invalid" }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ peerIdentity: toStringThrows }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "Error" ("Error")
FAIL new RTCPeerConnection({ certificates: null }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ certificates: undefined }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ certificates: [] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ certificates: [null] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ certificates: [undefined] }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ iceCandidatePoolSize: toNumberThrows }) assert_throws: function "function () {
        eval(expr);
      }" threw object "ReferenceError: RTCPeerConnection is not defined" ("ReferenceError") expected object "TypeError" ("TypeError")
FAIL new RTCPeerConnection({ certificates: [certificate] }) RTCPeerConnection is not defined
FAIL new RTCPeerConnection({ certificates: [expiredCertificate] }) RTCPeerConnection is not defined
FAIL localDescription initial value RTCPeerConnection is not defined
FAIL currentLocalDescription initial value RTCPeerConnection is not defined
FAIL pendingLocalDescription initial value RTCPeerConnection is not defined
FAIL remoteDescription initial value RTCPeerConnection is not defined
FAIL currentRemoteDescription initial value RTCPeerConnection is not defined
FAIL pendingRemoteDescription initial value RTCPeerConnection is not defined
FAIL signalingState initial value RTCPeerConnection is not defined
FAIL iceGatheringState initial value RTCPeerConnection is not defined
FAIL iceConnectionState initial value RTCPeerConnection is not defined
FAIL connectionState initial value RTCPeerConnection is not defined
FAIL canTrickleIceCandidates initial value RTCPeerConnection is not defined
Harness: the test ran to completion.
